1 scanner + checkinputlength
changed to pass scanner to functions. scanner is called in. checkinputlength to make sure varchar length in database isn't violated for strings. still need to do it for year int but kind of don't care because you could put like 4000 or something idk. plus what if we're in the year 4000. idk if i wanna put that kind of constraint. who knows what the future holds.